+++ modules/validators/builtin/cpp/Gateway.cpp
@@ -45,6 +45,7 @@
 #include "mustBeNonNanBuiltin.hpp"
 #include "mustBeNonZeroBuiltin.hpp"
 #include "mustBeNonSparseBuiltin.hpp"
+#include "mustBeRealBuiltin.hpp"
 //=============================================================================
 using namespace Nelson;
 //=============================================================================
@@ -79,7 +80,10 @@ static const nlsGateway gateway[] = {
         CPP_BUILTIN },
     { "mustBeNonNan", (void*)Nelson::ValidatorsGateway::mustBeNonNanBuiltin, 0, -1, CPP_BUILTIN },
     { "mustBeNonZero", (void*)Nelson::ValidatorsGateway::mustBeNonZeroBuiltin, 0, -1, CPP_BUILTIN },
-    { "mustBeNonSparse", (void*)Nelson::ValidatorsGateway::mustBeNonSparseBuiltin, 0, -1, CPP_BUILTIN },
+    { "mustBeNonSparse", (void*)Nelson::ValidatorsGateway::mustBeNonSparseBuiltin, 0, -1,
+        CPP_BUILTIN },
+    { "mustBeReal", (void*)Nelson::ValidatorsGateway::mustBeRealBuiltin, 0, -1, CPP_BUILTIN },
+
 };
 //=============================================================================
 static bool